+++ packages/nitro-protocol/contracts/test/TESTNitroAdjudicator.sol
@@ -53,8 +53,8 @@ contract TESTNitroAdjudicator is NitroAdjudicator {
     function acceptableWhoSignedWhat(
         uint8[] memory whoSignedWhat,
         uint48 largestTurnNum,
-        uint256 nParticipants,
-        uint256 nStates
+        uint8 nParticipants,
+        uint8 nStates
     ) public pure returns (bool) {
         return _acceptableWhoSignedWhat(whoSignedWhat, largestTurnNum, nParticipants, nStates);
     }
